feat/input: Shorten component names
