+++ components/src/chip/chip.scss
