+++ components/src/close-button/close-button.scss
