+++ components/src/dialog/index.ts
@@ -1,1 +1,1 @@
-export * from './dialog-element';
+export * from './dialog';
