--- components/src/scrim/scrim-element.ts
@@ -1,17 +1,0 @@
-import { LitElement, unsafeCSS } from 'lit';
-import { customElement } from 'lit/decorators.js';
-
-import styleScrim from './scrim-element.scss';
-
-const AOTW_SCRIM = 'aotw-scrim';
-
-@customElement(AOTW_SCRIM)
-export class ScrimElement extends LitElement {
-  static styles = unsafeCSS(styleScrim);
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    AOTW_SCRIM: ScrimElement
-  }
-}
