+++ components/src/scrim/scrim.ts
@@ -1,0 +1,17 @@
+import { LitElement, unsafeCSS } from 'lit';
+import { customElement } from 'lit/decorators.js';
+
+import styleScrim from './scrim.scss';
+
+const AOTW_SCRIM = 'aotw-scrim';
+
+@customElement(AOTW_SCRIM)
+export class ScrimElement extends LitElement {
+  static styles = unsafeCSS(styleScrim);
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    AOTW_SCRIM: ScrimElement
+  }
+}
