+++ components/src/toggle-switch/toggle-switch.scss
