+++ components/src/toggle-switch/toggle-switch.ts
@@ -2,48 +2,43 @@ import { html, LitElement, unsafeCSS } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 
-import { Icon } from '../icon/icon';
+import styleToggleSwitch from './toggle-switch.scss';
 
-import styleToggleButton from './toggle-button-element.scss';
+const AOTW_TOGGLE_SWITCH = 'aotw-toggle-switch';
 
-const AOTW_TOGGLE_BUTTON = 'aotw-toggle-button';
-
-@customElement(AOTW_TOGGLE_BUTTON)
-export class ToggleButtonElement extends LitElement {
+@customElement(AOTW_TOGGLE_SWITCH)
+export class ToggleSwitchElement extends LitElement {
   @property({ type: Boolean })
   active = false;
 
   @property({ type: Boolean })
   disabled = false;
 
-  @property()
-  icon?: Icon;
+  @property({ type: String })
+  label?: string;
 
-  static styles = unsafeCSS(styleToggleButton);
+  static styles = unsafeCSS(styleToggleSwitch);
 
   render() {
     const classes = {
       active: this.active,
       disabled: this.disabled
-    };
-
-    const iconHTML = this.icon
-      ? html`<aotw-icon name=${this.icon}></aotw-icon>`
-      : null;
+    }
 
     return html`
-      <button
-        class="toggle-button ${classMap(classes)}"
+      <div
+        class="toggle-switch ${classMap(classes)}"
         @click=${() => this.toggleActive()}
       >
-        ${iconHTML}
-        <slot></slot>
-      </button>
+        <div class="switch"></div>
+        <div class="handle"></div>
+      </div>
+      <p>${this.label}</p>
     `;
   }
 
-  toggleActive() {
-    if (this.disabled === false) {
+  private toggleActive() {
+    if (!this.disabled) {
       const onClick = new CustomEvent<boolean>('onClick', {
         detail: (this.active = !this.active)
       });
@@ -54,6 +49,6 @@ export class ToggleButtonElement extends LitElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    AOTW_TOGGLE_BUTTON: ToggleButtonElement
+    AOTW_TOGGLE_SWITCH: ToggleSwitchElement
   }
 }
