+++ storybook/.storybook/preview.js
@@ -14,5 +14,5 @@ export const parameters = {
       color: /(background|color)$/i,
       date: /Date$/,
     },
-  },
+  }
 }
