+++ storybook/src/components/input/_overview.stories.mdx
@@ -1,0 +1,20 @@
+import { Canvas, Meta } from '@storybook/addon-docs';
+
+<Meta title="Components/Input/Overview" />
+
+# AOTW Input
+
+<aotw-input value="Input"></aotw-input>
+
+```html
+<aotw-input
+  .label=[label-text]
+  .message=[message-text/message-state]
+  .placeholder=[placeholder-text]
+  .value=[value-text]
+></aotw-input>
+```
+
+---
+
+&copy; Jurrit van der Ploeg, 2022
